+++ src/main/client/system/index.ts
@@ -4,5 +4,6 @@ import './consoleCommand.js';
 import './native.js';
 import './notification.js';
 import './stats.js';
+import './waypoint.js';
 
 alt.log('Loaded System');