+++ src/main/server/index.ts
@@ -4,17 +4,15 @@ import { useApi } from './api/index.js';
 
 import { useConfig } from './config/index.js';
 
-import { useBlipLocal } from './controllers/blip.js';
-import { useBlipGlobal } from './controllers/blip.js';
-import { usePickupGlobal } from './controllers/pickups.js';
+import { useBlipGlobal, useBlipLocal } from './controllers/blip.js';
 import { useInteraction } from './controllers/interaction.js';
-import { useObjectGlobal, useObjectLocal } from './controllers/object.js';
 import { useMarkerGlobal, useMarkerLocal } from './controllers/markers.js';
+import { useObjectGlobal, useObjectLocal } from './controllers/object.js';
+import { usePickupGlobal } from './controllers/pickups.js';
 import { useTextLabelGlobal, useTextLabelLocal } from './controllers/textlabel.js';
 
 import { useDatabase } from './database/index.js';
 
-import { CollectionNames } from './document/shared.js';
 import {
     useAccount,
     useAccountBinder,
@@ -22,39 +20,43 @@ import {
     useCharacter,
     useCharacterBinder,
     useGlobal,
-    useVehicle as useVehicleDocument,
     useVehicleBinder,
+    useVehicle as useVehicleDocument,
     useVehicleEvents,
     useVirtual,
 } from './document/index.js';
+import { CollectionNames } from './document/shared.js';
 
 import { useEvents } from './events/index.js';
 
-import { useWorldGetter } from './getters/world.js';
 import { usePlayerGetter } from './getters/player.js';
 import { usePlayersGetter } from './getters/players.js';
 import { useVehicleGetter } from './getters/vehicle.js';
 import { useVehiclesGetter } from './getters/vehicles.js';
+import { useWorldGetter } from './getters/world.js';
 
+import { useAnimation } from './player/animation.js';
+import { usePlayerAppearance } from './player/appearance.js';
 import { useAudio } from './player/audio.js';
-import { useWorld } from './player/world.js';
+import { useClothing } from './player/clothing.js';
 import { useNative } from './player/native.js';
 import { useNotify } from './player/notify.js';
 import { useStatus } from './player/status.js';
-import { useWebview } from './player/webview.js';
-import { useClothing } from './player/clothing.js';
-import { useAnimation } from './player/animation.js';
-import { usePlayerAppearance } from './player/appearance.js';
 import { useWeapon } from './player/weapon.js';
+import { useWebview } from './player/webview.js';
+import { useWorld } from './player/world.js';
 
 import { useMessenger } from './systems/messenger.js';
 import { usePermission } from './systems/permission.js';
 import { usePermissionGroup } from './systems/permissionGroup.js';
 
-import { check, hash } from './utility/password.js';
-import { sha256, sha256Random } from './utility/hash.js';
-import { useVehicle } from './vehicle/index.js';
+import { usePlayer } from './player/index.js';
 import { useState } from './player/state.js';
+import { useWaypoint } from './player/waypoint.js';
+import { sha256, sha256Random } from './utility/hash.js';
+import { check, hash } from './utility/password.js';
+import { useVehicle } from './vehicle/index.js';
+import { useProtectCallback } from './utility/protectCallback.js';
 
 export function useRebar() {
     return {
@@ -118,6 +120,7 @@ export function useRebar() {
             useNative,
             useNotify,
             useState,
+            useWaypoint,
             useWeapon,
             useWebview,
             useWorld,
@@ -129,6 +132,7 @@ export function useRebar() {
             usePermission,
             usePermissionGroup,
         },
+        usePlayer,
         utility: {
             sha256,
             sha256Random,
@@ -136,6 +140,7 @@ export function useRebar() {
                 check,
                 hash,
             },
+            useProtectCallback,
         },
         vehicle: {
             useVehicle,